Refactor

+++ Framework/src/main/java/de/rub/nds/x509anvil/framework/featureextraction/probe/VersionProbe.java
@@ -1,3 +1,12 @@
+/**
+ * Framework - A tool for creating arbitrary certificates
+ * <p>
+ * Copyright 2014-2024 Ruhr University Bochum, Paderborn University, Hackmanit GmbH
+ * <p>
+ * Licensed under Apache License, Version 2.0
+ * http://www.apache.org/licenses/LICENSE-2.0.txt
+ */
+
 package de.rub.nds.x509anvil.framework.featureextraction.probe;
 
 import de.rub.nds.x509anvil.framework.featureextraction.probe.result.ProbeResult;
@@ -5,6 +14,8 @@ import de.rub.nds.x509anvil.framework.featureextraction.probe.result.VersionProb
 import de.rub.nds.x509anvil.framework.verifier.VerifierResult;
 import de.rub.nds.x509anvil.framework.x509.config.X509CertificateConfigUtil;
 import de.rub.nds.x509anvil.framework.x509.config.X509CertificateChainConfig;
+
+import java.math.BigInteger;
 
 public class VersionProbe extends SimpleProbe {
     private final int version;
@@ -16,7 +27,7 @@ public class VersionProbe extends SimpleProbe {
     @Override
     public X509CertificateChainConfig prepareConfig() {
         X509CertificateChainConfig x509CertificateChainConfig = X509CertificateConfigUtil.createBasicConfig(2);
-        x509CertificateChainConfig.getEntityCertificateConfig().setVersion(version);
+        x509CertificateChainConfig.getEntityCertificateConfig().setVersion(BigInteger.valueOf(version));
         return x509CertificateChainConfig;
     }
